Added a kind of blogging system
I added some javascript to the [main page](http://labs.antobase.com). This script will grab all commits, check if there is a url referencing http://labs.antobase.com in them, and list them as blog posts. It's very simple, but I think this will be enough.

The posts will automatically be *marked up* and the link to the full commit you can find under this post, along with the date and the author.